hw(chimney): Improve and fix setting the `last` flag
setting the `last` flags causes the router to do wormhole routing
i.e. it blocks the arbitration to a certain direction until the entire stream has been sent,
which corresponds to the `last` flag being asserted.

This is necessary for W bursts which need to be in order. However,
for R bursts this is not necessary actually, since they are already ordered by the AXI Id (which the Ws do not have).

This commit also makes sure that the AW and the W's are not split up which could lead to interleaving of multiple write requests from other sources