hw(chimney): Join AW and W
AW and the W beats should not be separated and routed wormhole-style.
The chimney has a wormhole arbiter which means it will look the direction, until the
`last` flag is observed. This means that after an AW it would never change the direction,
since the AWs don't have the `last` flag set. The solution for this problem is to
mux them to the same input channel to the arbiter