chimney: Fix ordering of `AW` and `W` beats (#25)
* floogen(link): Change the mapping for wide aw

* hw(nw_chimney): Send the wide AW over the wide channel

This is needed for correct ordering

* hw(chimney): Improve and fix setting the `last` flag

setting the `last` flags causes the router to do wormhole routing
i.e. it blocks the arbitration to a certain direction until the entire stream has been sent,
which corresponds to the `last` flag being asserted.

This is necessary for W bursts which need to be in order. However,
for R bursts this is not necessary actually, since they are already ordered by the AXI Id (which the Ws do not have).

This commit also makes sure that the AW and the W's are not split up which could lead to interleaving of multiple write requests from other sources

* hw(pkg): Regenerate the narrow wide package

* hw(chimney): Replace the wormhole arbiters with normal rr_arb's

The wormhole functionality is not needed in the network interface itself.
Otherwise the AW and W will block, since AW does not set the `last` flag

* hw(chimney): Take the last signal from the axi message not the header

Both header and AXI message might have a `last` flag. The last flag in the header is used for wormhole-routing,
while the `last` flag in the AXI message is used for tracking outstanding requests and responses

* hw(chimney): Fix valid in signal

* sim(wave): Update wave scripts

* doc: Update CHANGELOG

* Revert "hw(chimney): Replace the wormhole arbiters with normal rr_arb's"

This reverts commit 4095791647a2cd9d975430e2c8db6a019bf5de65.

* hw(chimney): Join AW and W

AW and the W beats should not be separated and routed wormhole-style.
The chimney has a wormhole arbiter which means it will look the direction, until the
`last` flag is observed. This means that after an AW it would never change the direction,
since the AWs don't have the `last` flag set. The solution for this problem is to
mux them to the same input channel to the arbiter

* hw(chimney): Use correct hdr information for the meta buffer