hw: Remove wormhole arbiter causing deadlock in router
When an endpoint initiates a wide multicast DMA transfer from another
endpoint to itself (and possibly other endpoints), the following
deadlock occurs. The DMA issues an AR, causing a read burst to come
back from the router to the initiator endpoint. When the first read
beats arrive, the DMA issues a write burst. This write burst loops
back to the initiator endpoint, and may take control of the physical
link. As writes lock the link for the entire burst, but the burst
cannot complete as it needs the stalling read beats to feed the write
burst (a DMA requirement), there is a deadlock. Note, this can happen
also without multicast, so long as the system uses loopback.